fix: directives incorrectly placed before implements
The GraphQL spec is very clear about the order of directives in a definition - they come after implements. https://spec.graphql.org/October2021/#sec-Interfaces

It seems this is not strictly enforced by all parsers, such as graphql-codegen, which works fine when mis-ordered. Others like graphql-inspector are strict and error when mis-ordered.

This was already correct for Objects.